ai/requirements.txt
requests==2.31.0
transformers==4.36.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv==1.0.0
